pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "python-acad-tools"
version = "0.1.0"
description = "Python tools for processing GIS data and exporting to AutoCAD DXF format"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "Gunther Schulz" }
]
dependencies = [
    "ezdxf",
    "geopandas",
    "numpy",
    "shapely",
    "pyproj",
    "pyyaml",
    "owslib",
    "pandas",
    "networkx",
]

[project.scripts]
acad-tools = "src.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build]
exclude = [
    "projects/*",
    "logs/*",
    "*.log",
    "__pycache__",
    "*.pyc",
    "OLD_CODE/*",
]

[tool.hatch.build.targets.wheel.sources]
"src" = "src"
"src/extensions" = "src/extensions"  # Include extensions directory 